fix: modify middleware isAdmin

+++ routes/api.php
@@ -33,12 +33,12 @@ Route::delete('/users', [UserController::class, 'deleteUsers'] );
 Route::get('/pizzas', [PizzaController::class, 'getAllPizzas'] );
 Route::post('/pizzas', [PizzaController::class, 'createPizza'] );
 Route::put('/pizzas/{id}', [PizzaController::class, 'updatePizza'] );
-Route::delete('/pizzas/{id}', [PizzaController::class, 'deletePizza']);
+Route::delete('/pizzas/{id}', [PizzaController::class, 'deletePizza'])->middleware('isAdmin');
 Route::get('/pizzas/{id}', [PizzaController::class, 'getPizzaById']);
-
 Route::get('/pizzas/reviews/{id}', [PizzaController::class, 'getPizzaByIdWithReviews']);
 Route::get('/pizzas/ingredients/{id}', [PizzaController::class, 'getPizzaByIdWithIngredients']);
 Route::post('/pizzas/add-ingredient/{id}', [PizzaController::class, 'addIngredientToPizzaId']);
+
 
 // Ingredients
 Route::get('/ingredients', [IngredientController::class, 'getAllIngredients'] );
